requirements.txt
pystray>=0.19.0
Pillow>=8.0.0
PyGObject>=3.36.0
